README.md
Weather King is a weather forecasting application for android which provides the detailed weather forecast of different locations of the world quickly and easily. It provides the users with a classic and user friendly interface to make their life easy by providing the information with just a few clicks. It also provides the feature of capturing the user current location based on the user preferences. The app also provides a cool and good looking widget to keep the user updated with the current weather.